Add git abbreviations only when git installed

+++ conf.d/fishabbreviations.fish
@@ -6,12 +6,18 @@ function _fishabbreviations_install --on-event fishabbreviations_install
   abbr --add -- w 'while test'
   abbr --add -- f for
 
-  abbr --add -- gcl 'git clone'
-  abbr --add -- gfe 'git fetch'
-  abbr --add -- gin 'git init'
-  abbr --add -- gsw 'git switch'
-  abbr --add -- gsy 'git pull && git push'
-  abbr --add -- gun 'rm -rf .git'
+  if command --query git
+    abbr --add -- gcl 'git clone'
+    abbr --add -- gfe 'git fetch'
+    abbr --add -- gin 'git init'
+    abbr --add -- gsw 'git switch'
+    abbr --add -- gsy 'git pull && git push'
+    abbr --add -- gun 'rm -rf .git'
+  end
+end
+
+function _fishabbreviations_update --on-event fishabbreviations_update
+  _fishabbreviations_install
 end
 
 function _fishabbreviations_uninstall --on-event fishabbreviations_uninstall
